cloud/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-cloud-storage==2.18.0
google-cloud-tasks==2.16.0
google-auth==2.35.0
cloud-sql-python-connector[pg8000]>=1.12.0
firebase-admin==6.5.0
